Move things around, extract filtering to a separate folder
